1.南讯项目更新

+++ nascent-bridge/src/main/java/com/me/nascent/modules/order/entity/Order.java
@@ -84,7 +84,7 @@ public class Order {
     private Integer refundTimes;
 
     @TableField("goodsLibId")
-    private Integer goodsLibId;
+    private Long goodsLibId;
 
     @TableField("warehouseId")
     private Integer warehouseId;
